requirements.txt
streamlit
flask
flask-restx
python-dotenv
openai
langchain
langchain-openai
litellm
langfuse
homeassistant_api
chromadb
flake8
pylint
Jinja2
coloredlogs
